// src/Umbraco.Community.Forms.Workflows/Client/src/generated/types.gen.ts
export type BasicClient = {
    clientID?: string | null;
    name?: string | null;
};

export type BasicList = {
    listID?: string | null;
    name?: string | null;
};